refactor: #129 Geometry DTO 파일 분리
- CategoryPointGeomDTO를 별도 파일로 분리
- CategoryLineStringGeomDTO를 별도 파일로 분리
- CategoryItemListDTO에서 Geometry DTO 정의 제거
