+++ Projects/Client/Category/Sources/Client/CategoryClient+Live.swift
@@ -23,5 +23,6 @@ extension CategoryClient: DependencyKey {
         return try await apiClient.execute(endpoint).toEntity()
       }
     )
+    
   }
 }
