+++ Projects/Client/Category/Sources/DTO/Response/CategoryItemListDTO.swift
@@ -74,39 +74,3 @@ extension CategoryItemResponseDTO {
   }
 }
 
-struct CategoryPointGeomDTO: DTO {
-  var type: String
-  var coordinates: [Double]
-
-  func toEntity() throws -> Coordinate {
-    guard coordinates.count >= 2 else {
-      throw FoundationError.failedToDecode(CategoryPointGeomDTO.self)
-    }
-    return Coordinate(latitude: coordinates[1], longitude: coordinates[0])
-  }
-}
-
-struct CategoryLineStringGeomDTO: DTO {
-  typealias Entity = [Coordinate]?
-  var type: String
-  var coordinates: [[Double]]
-  
-  init(
-    type: String,
-    coordinates: [[Double]]
-  ) {
-    self.type = type
-    self.coordinates = coordinates
-  }
-}
-
-extension CategoryLineStringGeomDTO {
-  func toEntity() throws -> [Coordinate]? {
-    let points = coordinates.compactMap { coord -> Coordinate? in
-      guard coord.count == 2 else { return nil }
-      
-      return Coordinate(latitude: coord[1], longitude: coord[0])
-    }
-    return points.isEmpty ? nil : points
-  }
-}
